// src/shared/api/generated/models/caseIntelligenceCommandLookupResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { CaseIntelligenceCommandLookupResponseDtoResult } from "./caseIntelligenceCommandLookupResponseDtoResult";
import type { CaseIntelligenceCommandLookupResponseDtoResultKind } from "./caseIntelligenceCommandLookupResponseDtoResultKind";

export interface CaseIntelligenceCommandLookupResponseDto {
  createdAt: string;
  /** @maxLength 64 */
  operation: string;
  result: CaseIntelligenceCommandLookupResponseDtoResult;
  resultKind: CaseIntelligenceCommandLookupResponseDtoResultKind;
}