// src/shared/api/generated/models/caseIntelligenceDetectionCompileResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { CompiledCaseIntelligenceDetectionPolicyDto } from "./compiledCaseIntelligenceDetectionPolicyDto";
import type { CaseIntelligenceDetectionCompileResponseDtoCompilerRevisionId } from "./caseIntelligenceDetectionCompileResponseDtoCompilerRevisionId";
import type { CaseIntelligenceDetectionCompileResponseDtoRouterInput } from "./caseIntelligenceDetectionCompileResponseDtoRouterInput";

export interface CaseIntelligenceDetectionCompileResponseDto {
  compiledPolicy: CompiledCaseIntelligenceDetectionPolicyDto;
  /**
   * @minLength 64
   * @maxLength 64
   */
  compiledPolicyHash: string;
  compilerRevisionId: CaseIntelligenceDetectionCompileResponseDtoCompilerRevisionId;
  routerInput?: CaseIntelligenceDetectionCompileResponseDtoRouterInput;
}