// src/shared/api/generated/models/caseIntelligenceDetectionRevisionResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { CompiledCaseIntelligenceDetectionPolicyDto } from "./compiledCaseIntelligenceDetectionPolicyDto";
import type { CaseIntelligenceDetectionRevisionResponseDtoCreatedByCmsUserId } from "./caseIntelligenceDetectionRevisionResponseDtoCreatedByCmsUserId";
import type { CaseIntelligenceDetectionPolicyDto } from "./caseIntelligenceDetectionPolicyDto";
import type { CaseIntelligenceDetectionRevisionResponseDtoPublishedAt } from "./caseIntelligenceDetectionRevisionResponseDtoPublishedAt";
import type { CaseIntelligenceDetectionRevisionResponseDtoPublishedByCmsUserId } from "./caseIntelligenceDetectionRevisionResponseDtoPublishedByCmsUserId";
import type { CaseIntelligenceDetectionRevisionResponseDtoStatus } from "./caseIntelligenceDetectionRevisionResponseDtoStatus";

export interface CaseIntelligenceDetectionRevisionResponseDto {
  compiledPolicy: CompiledCaseIntelligenceDetectionPolicyDto;
  /**
   * @minLength 64
   * @maxLength 64
   */
  compiledPolicyHash: string;
  /** @maxLength 128 */
  compilerRevisionId: string;
  createdAt: string;
  /** @nullable */
  createdByCmsUserId?: CaseIntelligenceDetectionRevisionResponseDtoCreatedByCmsUserId;
  definition: CaseIntelligenceDetectionPolicyDto;
  id: string;
  projectId: string;
  /** @nullable */
  publishedAt?: CaseIntelligenceDetectionRevisionResponseDtoPublishedAt;
  /** @nullable */
  publishedByCmsUserId?: CaseIntelligenceDetectionRevisionResponseDtoPublishedByCmsUserId;
  status: CaseIntelligenceDetectionRevisionResponseDtoStatus;
  /** @minimum 1 */
  version: number;
}